+++ Project_K/Object/Player/Elf/Elf.cpp
@@ -175,10 +175,10 @@ void Elf::Draw()
 
 #if _DEBUG
 	// プレイヤーのサイズ
-	DrawBox(m_sizeLeft + m_pos.x , 
-			m_sizeTop + m_pos.y,
-			m_sizeRight + m_pos.x, 
-			m_sizeBottom + m_pos.y,
+	DrawBox(m_sizeLeft + static_cast<int>(m_pos.x) ,
+			m_sizeTop + static_cast<int>(m_pos.y),
+			m_sizeRight + static_cast<int>(m_pos.x),
+			m_sizeBottom + static_cast<int>(m_pos.y),
 			0xffffff, false);
 	// m_attackFlagがtrueのとき攻撃当たり判定を表示
 	if (m_attackFlag)
